Move build_conv2d_connection to builder
